requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
bech32==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cosmpy==0.11.2
distlib==0.4.0
ecdsa==0.19.1
filelock==3.20.3
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
multidict==6.7.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
sortedcontainers==2.4.0
squareup==43.2.0.20251016
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uagents==0.23.5
uagents-core==0.4.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
virtualenv==20.36.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
bech32==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cosmpy==0.11.2
distlib==0.4.0
distro==1.9.0
ecdsa==0.19.1
filelock==3.20.3
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
multidict==6.7.0
google-genai
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
squareup==43.2.0.20251016
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uagents==0.23.5
uagents-core==0.4.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
virtualenv==20.36.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
bech32==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cosmpy==0.11.2
distlib==0.4.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-checkpoint-postgres==3.0.3
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
multidict==6.7.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
squareup==43.2.0.20251016
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uagents==0.23.5
uagents-core==0.4.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
virtualenv==20.36.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
